scheduler/requirements.txt
apache-airflow==2.10.2
pendulum==3.0.0

